backend/pyproject.toml
[tool.poetry]
name = "family-expense-tracker-backend"
version = "0.1.0"
description = "Family Expense Tracker API"
authors = ["Sumana Chakraborty <chakrabortysumana067@outlook.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
aiomysql = "^0.2.0"
bcrypt = "^4.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pydantic-settings = "^2.2.0"
pydantic = { extras = ["email"], version = "^2.7.0" }
alembic = "^1.13.0"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
anyio = "^4.4.0"
aiosqlite = "^0.20.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
